* Agents Model Rework #1

+++ src/main/java/br/ufjf/pgcc/eseco/domain/dao/core/DeveloperDAO.java
@@ -5,7 +5,7 @@ import br.ufjf.pgcc.eseco.domain.model.core.Developer;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class DeveloperDAO extends CommonDAO<Developer, Long> {
+public class DeveloperDAO extends CommonDAO<Developer, Integer> {
 
     protected DeveloperDAO() {
         super(Developer.class);
