+++ src/main/java/br/ufjf/pgcc/eseco/domain/dao/core/InstitutionDAO.java
@@ -5,7 +5,7 @@ import br.ufjf.pgcc.eseco.domain.model.core.Institution;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class InstitutionDAO extends CommonDAO<Institution, Long> {
+public class InstitutionDAO extends CommonDAO<Institution, Integer> {
 
     protected InstitutionDAO() {
         super(Institution.class);
